fix: add utils folder to functions folder + update every import + clean code
