+++ src/class/Player.js
@@ -1,13 +1,4 @@
-function generateSlug(str) {
-    return str
-      .normalize("NFD")                      // Décompose les accents
-      .replace(/[\u0300-\u036f]/g, "")       // Supprime les diacritiques
-      .replace(/[^a-zA-Z0-9\s_]/g, "")       // Supprime les caractères spéciaux (mais garde underscore)
-      .trim()                                // Enlève les espaces inutiles
-      .replace(/\s+/g, "_")                  // Remplace les espaces par des underscores
-      .replace(/_+/g, "_")                   // Supprime les doubles underscores
-      .toLowerCase();
-}
+const { generateSlug } = require("../functions/utils/generateSlug");
 
 class Player {
     constructor(name, id = null, totalKills = null, totalDeaths = null) {
