+++ src/functions/utils/generateSlug.js
@@ -1,0 +1,12 @@
+function generateSlug(str) {
+    return str
+      .normalize("NFD")                      // Décompose les accents
+      .replace(/[\u0300-\u036f]/g, "")       // Supprime les diacritiques
+      .replace(/[^a-zA-Z0-9\s_]/g, "")       // Supprime les caractères spéciaux (mais garde underscore)
+      .trim()                                // Enlève les espaces inutiles
+      .replace(/\s+/g, "_")                  // Remplace les espaces par des underscores
+      .replace(/_+/g, "_")                   // Supprime les doubles underscores
+      .toLowerCase();
+}
+
+module.exports = { generateSlug };
